chore: exclude schemas from code duplication analysis
Schemas are best read as a block in one place instead of being broken down into functions
purely for the purpose of code reuse. The string values will be unique for all schema elements
so the savings from code reuse will be a couple of lines for each element that defines
the type and whether or not the element is nullable, which are easier to read when they
are clearly set in the schema definition block.